Move story play screen to a module
