+++ lib/modules/store_play/screens/play.dart
@@ -5,9 +5,9 @@ import 'package:andax/models/node.dart';
 import 'package:andax/models/story.dart';
 import 'package:andax/models/transition.dart';
 import 'package:andax/models/translation_asset.dart';
-import 'package:andax/utils.dart';
-import 'package:andax/widgets/happiness_slider.dart';
-import 'package:andax/widgets/node_card.dart';
+import '../utils/get_translation.dart';
+import 'package:andax/modules/store_play/widgets/happiness_slider.dart';
+import 'package:andax/modules/store_play/widgets/node_card.dart';
 import 'package:flutter/material.dart';
 import 'package:simple_animations/simple_animations.dart';
