+++ lib/modules/store_play/widgets/happiness_slider.dart
