+++ lib/modules/store_play/widgets/node_card.dart
@@ -1,7 +1,7 @@
 import 'package:andax/models/actor.dart';
 import 'package:andax/models/node.dart';
 import 'package:andax/models/translation_asset.dart';
-import 'package:andax/utils.dart';
+import '../utils/get_translation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_markdown/flutter_markdown.dart';
