+++ lib/utils.dart
@@ -1,5 +1,3 @@
-import 'models/translation_asset.dart';
-
 List<T> listFromJson<T>(
   Object? array,
   T Function(dynamic) fromJson,
@@ -10,15 +8,6 @@ List<String>? json2list(Object? array) {
   return (array as Iterable<dynamic>?)
       ?.map((dynamic i) => i as String)
       .toList();
-}
-
-String getTranslation<T extends TranslationAsset>(
-  Map<String, TranslationAsset> translations,
-  String id,
-  String? Function(T) getter,
-) {
-  final asset = translations[id] as T?;
-  return asset == null ? '' : getter(asset) ?? '';
 }
 
 String capitalize(String? text) {
